+++ src/post-processing/gifski-post-processor.vala
@@ -24,41 +24,10 @@ namespace Peek.PostProcessing {
       this.quality = quality;
     }
 
-    public async File? process_async (File file) {
-        var status = yield generate_pngs_async (file);
-
-        if (!Utils.is_exit_status_success (status)) {
-          return null;
-        }
-
-        var output_file = yield generate_animation_async (file);
-        return output_file;
-    }
-
-    public void cancel () {
-      if (pid != null) {
-        Posix.kill (pid, Posix.SIGINT);
-      }
-    }
-
-    private async int generate_pngs_async (File file) {
-      string[] args = {
-        "ffmpeg", "-y",
-        "-i", file.get_path (),
-        get_png_filename_pattern (file, "%04d")
-      };
-
-      return yield spawn_file_conversion_async (args);
-    }
-
-    private async File? generate_animation_async (File input_file) {
+    public async File[]? process_async (File[] files) {
       try {
         var extension = Utils.get_file_extension_for_format (OUTPUT_FORMAT_GIF);
         var output_file = Utils.create_temp_file (extension);
-
-        var png_file_pattern = get_png_filename_pattern (input_file, "*");
-        var glob = Posix.Glob ();
-        glob.glob (png_file_pattern);
 
         string[] args = {
           "gifski",
@@ -69,7 +38,9 @@ namespace Peek.PostProcessing {
 
         var argv = new Array<string> ();
         argv.append_vals (args, args.length);
-        argv.append_vals (glob.pathv, (uint)glob.pathc);
+        foreach (var file in files) {
+          argv.append_val (file.get_path ());
+        }
 
         int status = yield spawn_file_conversion_async (argv.data);
 
@@ -77,14 +48,16 @@ namespace Peek.PostProcessing {
           FileUtils.remove (output_file);
         }
 
-        foreach (string temp_file_path in glob.pathv) {
-          FileUtils.remove (temp_file_path);
-        }
-
-        return File.new_for_path (output_file);
+        return { File.new_for_path (output_file) };
       } catch (FileError e) {
         stderr.printf ("Error: %s\n", e.message);
         return null;
+      }
+    }
+
+    public void cancel () {
+      if (pid != null) {
+        Posix.kill (pid, Posix.SIGINT);
       }
     }
 
@@ -114,10 +87,6 @@ namespace Peek.PostProcessing {
     private void spawn (string[] argv) throws SpawnError {
       Process.spawn_async (null, argv, null,
         SpawnFlags.SEARCH_PATH | SpawnFlags.DO_NOT_REAP_CHILD, null, out pid);
-    }
-
-    private static string get_png_filename_pattern (File input_file, string replacement) {
-      return input_file.get_path () + "." + replacement + ".png";
     }
   }
 }